fix(ci): repair cross-platform test target builds
