+++ tools/compare_snapshots/CMakeLists.txt
@@ -20,7 +20,10 @@ if(ZEROSPADES_TESTS)
     # FP-determinism flags (same as zerospades_tests / zerospades_testable)
     if(CMAKE_CXX_COMPILER_ID MATCHES "GNU|Clang")
         target_compile_options(compare_snapshots PRIVATE
-            -fno-fast-math -ffp-contract=off -msse2 -mfpmath=sse)
+            -fno-fast-math -ffp-contract=off)
+        if(ZEROSPADES_USE_SSE_FP_FLAGS)
+            target_compile_options(compare_snapshots PRIVATE -msse2 -mfpmath=sse)
+        endif()
     elseif(MSVC)
         target_compile_options(compare_snapshots PRIVATE /fp:strict)
     endif()
